# load-testing-swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Anecole Backend API - Load Testing",
    "description": "API спецификация для нагрузочного тестирования образовательной платформы Anecole",
    "version": "1.0.0",
    "contact": {
      "name": "Anecole Development Team",
      "email": "dev@anecole.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local Development Server"
    },
    {
      "url": "http://10.10.3.4:8080/api",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "student1"
          },
          "email": {
            "type": "string",
            "example": "student1@school.com"
          },
          "firstName": {
            "type": "string",
            "example": "Алексей"
          },
          "lastName": {
            "type": "string",
            "example": "Иванов"
          },
          "role": {
            "type": "string",
            "enum": ["STUDENT", "TEACHER", "ADMIN"],
            "example": "STUDENT"
          },
          "grade": {
            "type": "integer",
            "example": 5
          },
          "school": {
            "type": "string",
            "example": "Школа №1"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/users/me": {
      "get": {
        "tags": ["Пользователи"],
        "summary": "Получить текущего пользователя",
        "description": "Получение информации о текущем аутентифицированном пользователе",
        "operationId": "getCurrentUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о пользователе",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          }
        }
      }
    },
...
  },
...
} 